+++ packages/side-nav/src/SideNav.tsx
@@ -12,14 +12,31 @@ import {
   prefersReducedMotion,
   validateAriaLabelProps,
 } from '@leafygreen-ui/a11y';
+<<<<<<< HEAD
+=======
+import { useBaseFontSize } from '@leafygreen-ui/leafygreen-provider';
+import { useEventListener } from '@leafygreen-ui/hooks';
+import { uiColors } from '@leafygreen-ui/palette';
+import { css, cx } from '@leafygreen-ui/emotion';
+import { spacing } from '@leafygreen-ui/tokens';
+import { keyMap, IdAllocator, createDataProp } from '@leafygreen-ui/lib';
+>>>>>>> origin
 import { useUsingKeyboardContext } from '@leafygreen-ui/leafygreen-provider';
 import { sideNavWidth, ulStyleOverrides, collapseDuration } from './styles';
 import SideNavContext from './SideNavContext';
 import CollapseToggle from './CollapseToggle';
 
+<<<<<<< HEAD
+=======
+const navIdAllocator = IdAllocator.create('input');
+const dataProp = createDataProp('side-nav');
+const sideNavSelector = dataProp.selector;
+
+export { sideNavSelector };
+
+>>>>>>> origin
 const navStyles = css`
   transition: all ${collapseDuration}ms ease-in-out;
-  width: ${sideNavWidth}px;
   background-color: ${uiColors.gray.light3};
   border-right: 1px solid ${uiColors.gray.light2};
   position: relative;
@@ -66,10 +83,6 @@ const listStyles = css`
   bottom: 0;
 `;
 
-const expandedListStyle = css`
-  width: ${sideNavWidth}px;
-`;
-
 const wrapper = css`
   position: absolute;
   top: 0;
@@ -80,7 +93,6 @@ const wrapper = css`
 
 const space = css`
   transition: width ${collapseDuration}ms ease-in-out;
-  width: ${sideNavWidth}px;
   position: relative;
 
   ${prefersReducedMotion(`
@@ -140,6 +152,26 @@ interface SideNavProps {
   children?: React.ReactNode;
 
   id?: string;
+
+  /**
+   * Determines the base font size for the menu items.
+   */
+  baseFontSize?: 14 | 16;
+
+  /**
+   * Provides an override for the SideNav width.
+   */
+  widthOverride?: number;
+
+  /**
+   * Allows consuming applications to control the collapsed state of the navigation.
+   */
+  collapsed?: boolean;
+
+  /**
+   * Consuming application's collapsed-state management controller
+   */
+  setCollapsed?: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 /**
@@ -157,10 +189,23 @@ interface SideNavProps {
  *
  * @param props.className Class name that will be applied to the root-level element.
  * @param props.children Content that will be rendered inside the root-level element.
+ * @param props.baseFontSize Determines the base font size for the menu items.
+ * @param props.widthOverride Provides an override for the SideNav width.
+ * @param props.collapsed Allows consuming applications to control the collapsed state of the navigation.
+ * @param props.setCollapsed Consuming application's collapsed-state management controller
  */
-function SideNav({ className, children, id: idProp, ...rest }: SideNavProps) {
+function SideNav({
+  className,
+  children,
+  id: idProp,
+  baseFontSize,
+  widthOverride,
+  collapsed: controlledCollapsed,
+  setCollapsed: setControlledCollapsed = () => {},
+  ...rest
+}: SideNavProps) {
   const { Provider: ContextProvider } = SideNavContext;
-  const [collapsed, setCollapsed] = useState(false);
+  const [uncontrolledCollapsed, uncontrolledSetCollapsed] = useState(false);
   const [hover, setHover] = useState(false);
   const [focus, setFocus] = useState(false);
   const { usingKeyboard } = useUsingKeyboardContext();
@@ -169,6 +214,20 @@ function SideNav({ className, children, id: idProp, ...rest }: SideNavProps) {
     portalContainer,
     setPortalContainer,
   ] = useState<HTMLUListElement | null>(null);
+  const providerFontSize = useBaseFontSize();
+  const fontSize: 14 | 16 = baseFontSize ?? providerFontSize;
+  const width =
+    typeof widthOverride === 'number' ? widthOverride : sideNavWidth;
+
+  const collapsed =
+    typeof controlledCollapsed === 'boolean'
+      ? controlledCollapsed
+      : uncontrolledCollapsed;
+
+  const setCollapsed =
+    typeof controlledCollapsed === 'boolean'
+      ? setControlledCollapsed
+      : uncontrolledSetCollapsed;
 
   // We visually expand the navigation when a user focuses on an element within the navigation
   // while navigating via keyboard.
@@ -211,20 +270,38 @@ function SideNav({ className, children, id: idProp, ...rest }: SideNavProps) {
             navId,
             collapsed,
             portalContainer,
+            width,
             transitionState: state,
+            baseFontSize: fontSize,
           }}
         >
           <div
             data-testid="side-nav-container"
-            className={cx(space, { [collapsedSpace]: collapsed }, className)}
+            {...dataProp.prop}
+            className={cx(
+              space,
+              css`
+                width: ${width}px;
+              `,
+              { [collapsedSpace]: collapsed },
+              className,
+            )}
           >
             <div className={wrapper} onMouseLeave={() => setHover(false)}>
               <nav
                 id={navId}
-                className={cx(navStyles, {
-                  [collapsedNavStyles]: ['entering', 'entered'].includes(state),
-                  [hoverNavStyles]: (hover || focusExpand) && collapsed,
-                })}
+                className={cx(
+                  navStyles,
+                  css`
+                    width: ${width}px;
+                  `,
+                  {
+                    [collapsedNavStyles]: ['entering', 'entered'].includes(
+                      state,
+                    ),
+                    [hoverNavStyles]: (hover || focusExpand) && collapsed,
+                  },
+                )}
                 onFocus={() => setFocus(true)}
                 onBlur={() => setFocus(false)}
                 onMouseEnter={() => setHover(true)}
@@ -235,7 +312,9 @@ function SideNav({ className, children, id: idProp, ...rest }: SideNavProps) {
                     className={cx(
                       ulStyleOverrides,
                       listStyles,
-                      expandedListStyle,
+                      css`
+                        width: ${width}px;
+                      `,
                     )}
                   >
                     {children}