+++ packages/side-nav/src/SideNavGroup.tsx
@@ -11,11 +11,8 @@ import { css, cx } from '@leafygreen-ui/emotion';
 import { spacing } from '@leafygreen-ui/tokens';
 import { useIdAllocator } from '@leafygreen-ui/hooks';
 import CollapsedSideNavItem from './CollapsedSideNavItem';
-import {
-  ulStyleOverrides,
-  sideNavItemSidePadding,
-  sideNavWidth,
-} from './styles';
+import { ulStyleOverrides, sideNavItemSidePadding } from './styles';
+import { useSideNavContext } from './SideNavContext';
 
 const button = createDataProp('side-nav-group-button');
 
@@ -31,6 +28,7 @@ const labelStyle = css`
   align-items: center;
   justify-content: space-between;
   font-size: 12px;
+  line-height: 1em;
   letter-spacing: 0.3px;
   font-weight: bold;
   text-transform: uppercase;
@@ -39,7 +37,6 @@ const labelStyle = css`
   margin-top: 0;
   margin-bottom: 0;
   padding: 4px ${sideNavItemSidePadding}px 4px ${sideNavItemSidePadding}px;
-  line-height: 1em;
 
   &:not(:first-of-type) {
     margin-top: ${spacing[1]}px;
@@ -52,7 +49,6 @@ const collapsibleLabelStyle = css`
   margin: 0px;
   transition: border-color 150ms ease-in-out, color 150ms ease-in-out;
   cursor: pointer;
-  width: ${sideNavWidth}px;
   border-bottom: 1px solid ${uiColors.gray.light2};
 
   &:hover {
@@ -210,6 +206,7 @@ function SideNavGroup({
 
   const menuGroupLabelId = useIdAllocator({ prefix: 'menu-group-label-id' });
   const menuId = useIdAllocator({ prefix: 'menu' });
+  const { width } = useSideNavContext();
 
   const isActiveGroup: boolean = useMemo(() => {
     if (hasActiveItem != null) {
@@ -262,9 +259,16 @@ function SideNavGroup({
           {...button.prop}
           aria-controls={menuId}
           aria-expanded={open}
-          className={cx(labelStyle, collapsibleLabelStyle, {
-            [collapsibleHeaderFocusStyle]: showFocus,
-          })}
+          className={cx(
+            labelStyle,
+            collapsibleLabelStyle,
+            css`
+              width: ${width}px;
+            `,
+            {
+              [collapsibleHeaderFocusStyle]: showFocus,
+            },
+          )}
           onClick={() => setOpen(curr => !curr)}
           id={menuGroupLabelId}
           data-testid="side-nav-group-header-label"