+++ packages/leafygreen-provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leafygreen-ui/leafygreen-provider
 
+## 2.1.1
+
+### Patch Changes
+
+- 1fe3e1a3: Updates type of useBaseFontSize from number to `14 | 16` specifically.
+
 ## 2.1.0
 
 ### Minor Changes